pyproject.toml
[tool.poetry]
name = "agi"
version = "0.2.0"
description = "AGI: Toward General AI through Advanced Programming Language Understanding"
authors = ["Matthew <matthew@zeti.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tiktoken = "^0.5.2"
torch = "^2.1.1"
argparse = "^1.4.0"
lightning = "^2.1.2"
tensorboard = "^2.15.1"
pyyaml = "^6.0.1"
wandb = "^0.16.1"
google-cloud-storage = "^2.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
